rpc_server/requirements.txt
pika
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.1
scikit-learn==1.6.1
sqlalchemy==2.0.31
asyncpg==0.28.0
psycopg==3.1.10
psycopg-binary==3.2.4
alembic==1.12.0
sqlmodel==0.0.22
pydantic[email]==2.3.0
boto3
mlflow==2.22.0
cloudpickle==3.1.1
jinja2==3.1.6
numpy==2.2.5
pandas==2.2.3
scipy==1.15.2
torch==2.7.0
torch-geometric==2.6.1
torch-cluster==1.6.3
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2